(complib) Add dev server for component docs
