+++ components/Makefile
@@ -34,7 +34,9 @@ uninstall:
 # development
 #####################################################################
 
-
+.PHONY: dev
+dev:
+	$(env)=development $(bin)/styleguidist server
 
 # checks
 #####################################################################
