+++ components/package.json
@@ -53,6 +53,7 @@
     "babel-core": "^6.26.0",
     "babel-eslint": "8.0.3",
     "babel-jest": "^21.2.0",
+    "babel-loader": "^7.1.2",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
     "babel-preset-env": "^1.6.1",
     "babel-preset-flow": "^6.23.0",
@@ -63,12 +64,16 @@
     "flow-typed": "^2.2.3",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^21.2.1",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-styleguidist": "^6.1.0",
     "react-test-renderer": "^16.2.0",
     "snazzy": "^7.0.0",
     "standard": "^10.0.3",
     "stylelint": "^8.3.1",
     "stylelint-config-css-modules": "^1.1.0",
-    "stylelint-config-standard": "^18.0.0"
+    "stylelint-config-standard": "^18.0.0",
+    "webpack": "^3.10.0"
   },
   "peerDependencies": {
     "react": "^16.2.0",
