+++ components/src/buttons/buttons.css
@@ -2,6 +2,7 @@
 
 .button {
   padding: 0.5rem 1.5rem;
+  border: none;
   text-transform: uppercase;
 }
