+++ components/src/buttons/index.js
@@ -1,0 +1,8 @@
+// @flow
+// button components
+
+import Button from './Button'
+import PrimaryButton from './PrimaryButton'
+
+export * from './Button'
+export {Button, PrimaryButton}
