+++ components/src/icons/Icon.js
@@ -1,6 +1,4 @@
 // @flow
-// SVG icon component
-
 import * as React from 'react'
 import classnames from 'classnames'
 
@@ -8,15 +6,30 @@ import ICON_DATA_BY_NAME, {type IconName} from './icon-data'
 import styles from './icons.css'
 
 type Props = {
+  /** name constant of the icon to display */
   name: IconName,
+  /** classes to apply */
   className?: string,
+  /** spin the icon with a CSS animation */
   spin?: boolean,
+  /** x attribute as a number or string (for nesting inside another SVG) */
   x?: number | string,
+  /** y attribute as a number or string (for nesting inside another SVG) */
   y?: number | string,
+  /** width as a number or string (for nesting inside another SVG) */
   height?: number | string,
+  /** height as a number or string (for nesting inside another SVG) */
   width?: number | string
 }
 
+/**
+ * Inline SVG icon component
+ *
+ * If you need access to the IconName type, you can:
+ * ```js
+ * import {type IconName} from '@opentrons/components'
+ * ```
+ */
 export default function Icon (props: Props) {
   const {x, y, height, width} = props
   const {viewBox, path} = ICON_DATA_BY_NAME[props.name]
